Made code more clean
