--- frontend/src/elm/URL.elm
@@ -1,4 +1,0 @@
-module URL exposing (..)
-
-host : String
-host = "http://localhost:8000/"
