+++ frontend/src/static/index.js
@@ -1,2 +1,4 @@
 Elm = require('../elm/main');
 Elm.Main.embed(document.getElementById('main'));
+
+fileRequest = require('../js/filerequest');
